+++ src/components/GoToComicDialog/ComicList/GoToComicButton/GoToComicButton.tsx
@@ -32,6 +32,17 @@ export default function GoToComicButton({
             ) : (
                 <></>
             )}
+            {comic.isGuestComic ? (
+                <span className="bg-qc-header text-white font-bold text-sm rounded-full px-2 py-0.5 ml-2">
+                    Guest Comic
+                </span>
+            ) : comic.isNonCanon ? (
+                <span className="bg-qc-non-canon text-white font-bold text-sm rounded-full px-2 py-0.5 ml-2">
+                    Non-canon
+                </span>
+            ) : (
+                ''
+            )}
         </button>
     )
 }